Use the background job to sync simple products from bulk action

+++ facebook-commerce.php
@@ -614,44 +614,29 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 		if ( $fb_product_group_id ) {
 
 			?>
-				<?php echo esc_html__( 'Facebook ID:', 'facebook-for-woocommerce' ); ?>
-				<a href="https://facebook.com/<?php echo esc_attr( $fb_product_group_id ); ?>"
-				target="_blank">
-					<?php echo esc_html( $fb_product_group_id ); ?>
-				</a>
-				<p/>
-			<?php
 
-			if ( WC_Facebookcommerce_Utils::is_variable_type( $woo_product->get_type() ) ) {
+			<?php echo esc_html__( 'Facebook ID:', 'facebook-for-woocommerce' ); ?> <a href="https://facebook.com/<?php echo esc_attr( $fb_product_group_id ); ?>"
+			                                                                           target="_blank"><?php echo esc_html( $fb_product_group_id ); ?></a>
 
-				?>
-					<p><?php echo esc_html__( 'Variant IDs:', 'facebook-for-woocommerce' ); ?><br/>
-				<?php
+			<?php if ( WC_Facebookcommerce_Utils::is_variable_type( $woo_product->get_type() ) ) : ?>
 
-				$children = $woo_product->get_children();
+				<?php if ( $product_item_ids_by_variation_id = $this->get_variation_product_item_ids( $woo_product, $fb_product_group_id ) ) : ?>
 
-				foreach ( $children as $child_id ) {
+					<p>
+						<?php echo esc_html__( 'Variant IDs:', 'facebook-for-woocommerce' ); ?><br/>
 
-					$fb_product_item_id = $this->get_product_fbid(
-						self::FB_PRODUCT_ITEM_ID,
-						$child_id
-					);
+						<?php foreach ( $product_item_ids_by_variation_id as $variation_id => $product_item_id ) : ?>
 
-					?>
-						<?php echo esc_html( $child_id ); ?>:
-						<a href="https://facebook.com/<?php echo esc_attr( $fb_product_item_id ); ?>"
-						target="_blank">
-							<?php echo esc_html( $fb_product_item_id ); ?>
-						</a><br/>
-					<?php
-				}
+							<?php echo esc_html( $variation_id ); ?>: <a href="https://facebook.com/<?php echo esc_attr( $product_item_id ); ?>"
+							                                             target="_blank"><?php echo esc_html( $product_item_id ); ?></a><br/>
 
-				?>
+						<?php endforeach; ?>
 					</p>
-				<?php
-			}
 
-			?>
+				<?php endif; ?>
+
+			<?php endif; ?>
+
 				<?php /* ?>
 				<?php echo esc_html__( 'Visible:', 'facebook-for-woocommerce' ); ?>
 				<input name="<?php echo esc_attr( Products::VISIBILITY_META_KEY ); ?>"
@@ -685,6 +670,94 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 		?>
 			</span>
 		<?php
+	}
+
+
+	/**
+	 * Gets a list of Product Item IDs indexed by the ID of the variation.
+	 *
+	 * @since 2.0.0-dev.1
+	 *
+	 * @param string $product_group_id product group ID
+	 * @return array
+	 */
+	private function get_variation_product_item_ids( $product, $product_group_id ) {
+
+		$product_item_ids_by_variation_id = [];
+		$missing_product_item_ids         = [];
+
+		// get the product item IDs from meta data and build a list of variations that don't have a product item ID stored
+		foreach ( $product->get_children() as $variation_id ) {
+
+			if ( $variation = wc_get_product( $variation_id ) ) {
+
+				if ( $product_item_id = $variation->get_meta( self::FB_PRODUCT_ITEM_ID ) ) {
+
+					$product_item_ids_by_variation_id[ $variation_id ] = $product_item_id;
+
+				} else {
+
+					$retailer_id = WC_Facebookcommerce_Utils::get_fb_retailer_id( $variation );
+
+					$missing_product_item_ids[ $retailer_id ] = $variation;
+
+					$product_item_ids_by_variation_id[ $variation_id ] = null;
+				}
+			}
+		}
+
+		// use the Graph API to try to find and store the product item IDs for variations that don't have a value yet
+		if ( $missing_product_item_ids ) {
+
+			$product_item_ids = $this->find_variation_product_item_ids( $product_group_id );
+
+			foreach ( $missing_product_item_ids as $retailer_id => $variation ) {
+
+				if ( isset( $product_item_ids[ $retailer_id ] ) ) {
+
+					$variation->update_meta_data( self::FB_PRODUCT_ITEM_ID, $product_item_ids[ $retailer_id ] );
+					$variation->save_meta_data();
+
+					$product_item_ids_by_variation_id[ $variation->get_id() ] = $product_item_ids[ $retailer_id ];
+				}
+			}
+		}
+
+		return $product_item_ids_by_variation_id;
+	}
+
+
+	/**
+	 * Uses the Graph API to return a list of Product Item IDs indexed by the variation's retailer ID.
+	 *
+	 * @since 2.0.0-dev.1
+	 *
+	 * @param string $product_group_id product group ID
+	 * @return array
+	 */
+	private function find_variation_product_item_ids( $product_group_id ) {
+
+		$product_item_ids = [];
+
+		try {
+
+			$response = facebook_for_woocommerce()->get_api()->get_product_group_products( $product_group_id );
+
+			do {
+
+				$product_item_ids = array_merge( $product_item_ids, $response->get_ids() );
+
+			// get up to two additional pages of results
+			} while ( $response = facebook_for_woocommerce()->get_api()->next( $response, 2 ) );
+
+		} catch ( Framework\SV_WC_API_Exception $e ) {
+
+			$message = sprintf( 'There was an error trying to find the IDs for Product Items in the Product Group %s: %s', $product_group_id, $e->getMessage() );
+
+			facebook_for_woocommerce()->log( $message );
+		}
+
+		return $product_item_ids;
 	}
 
 
@@ -1025,30 +1098,25 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 	 * Will determine product type (simple or variable) and delegate to
 	 * appropriate handler.
 	 *
-	 * @param int $wp_id product ID
+	 * @param int $product_id product ID
 	 */
-	public function on_product_publish( $wp_id ) {
+	public function on_product_publish( $product_id ) {
 
 		// bail if the plugin is not configured properly
 		if ( ! $this->is_configured() || ! $this->get_product_catalog_id() ) {
 			return;
 		}
 
-		if ( get_post_status( $wp_id ) !== 'publish' ) {
+		$product = wc_get_product( $product_id );
+
+		if ( ! $this->product_should_be_synced( $product ) ) {
 			return;
 		}
 
-		$woo_product = new WC_Facebook_Product( $wp_id );
-
-		// skip if not enabled for sync
-		if ( ! $woo_product->woo_product instanceof \WC_Product || ! Products::product_should_be_synced( $woo_product->woo_product ) ) {
-			return;
-		}
-
-		if ( $woo_product->woo_product->is_type( 'variable' ) ) {
-			$this->on_variable_product_publish( $wp_id, $woo_product );
+		if ( $product->is_type( 'variable' ) ) {
+			$this->on_variable_product_publish( $product_id );
 		} else {
-			$this->on_simple_product_publish( $wp_id, $woo_product );
+			$this->on_simple_product_publish( $product_id );
 		}
 	}
 
@@ -1058,11 +1126,13 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 	 * this function will delete the product from FB Page as well.
 	 */
 	function delete_on_out_of_stock( $wp_id, $woo_product ) {
-		if ( get_option( 'woocommerce_hide_out_of_stock_items' ) === 'yes' &&
-		! $woo_product->is_in_stock() ) {
+
+		if ( Products::product_should_be_deleted( $woo_product ) ) {
+
 			$this->delete_product_item( $wp_id );
 			return true;
 		}
+
 		return false;
 	}
 
@@ -1075,24 +1145,20 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 	 */
 	function on_variable_product_publish( $wp_id, $woo_product = null ) {
 
-		if ( ! $this->is_product_sync_enabled() ) {
+		if ( ! $woo_product instanceof \WC_Facebook_Product ) {
+			$woo_product = new \WC_Facebook_Product( $wp_id );
+		}
+
+		if ( ! $this->product_should_be_synced( $woo_product->woo_product ) ) {
 			return;
 		}
 
-		if ( get_post_status( $wp_id ) != 'publish' ) {
+		if ( $this->delete_on_out_of_stock( $wp_id, $woo_product->woo_product ) ) {
 			return;
 		}
 
 		// Check if product group has been published to FB.  If not, it's new.
 		// If yes, loop through variants and see if product items are published.
-		if ( ! $woo_product ) {
-			$woo_product = new WC_Facebook_Product( $wp_id );
-		}
-
-		if ( $this->delete_on_out_of_stock( $wp_id, $woo_product ) ) {
-			return;
-		}
-
 		$fb_product_group_id = $this->get_product_fbid( self::FB_PRODUCT_GROUP_ID, $wp_id, $woo_product );
 
 		if ( $fb_product_group_id ) {
@@ -1128,24 +1194,15 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 	 */
 	function on_simple_product_publish( $wp_id, $woo_product = null, &$parent_product = null ) {
 
-		if ( ! $this->is_product_sync_enabled() ) {
+		if ( ! $woo_product instanceof \WC_Facebook_Product ) {
+			$woo_product = new \WC_Facebook_Product( $wp_id, $parent_product );
+		}
+
+		if ( ! $this->product_should_be_synced( $woo_product->woo_product ) ) {
 			return;
 		}
 
-		if ( get_post_status( $wp_id ) != 'publish' ) {
-			return;
-		}
-
-		if ( ! $woo_product ) {
-			$woo_product = new WC_Facebook_Product( $wp_id, $parent_product );
-		}
-
-		// skip if not enabled for sync
-		if ( ! $woo_product->woo_product instanceof \WC_Product || ! Products::product_should_be_synced( $woo_product->woo_product ) ) {
-			return;
-		}
-
-		if ( $this->delete_on_out_of_stock( $wp_id, $woo_product ) ) {
+		if ( $this->delete_on_out_of_stock( $wp_id, $woo_product->woo_product ) ) {
 			return;
 		}
 
@@ -1192,6 +1249,42 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 				return $this->create_product_simple( $woo_product );  // new product
 			}
 		}
+	}
+
+
+	/**
+	 * Determines whether the product with the given ID should be synced.
+	 *
+	 * TODO: can we move this logic into Products::product_should_be_synced()? {WV 2020-05-22}
+	 *
+	 * @since 2.0.0-dev.1
+	 *
+	 * @param \WC_Product|false $product product object
+	 */
+	public function product_should_be_synced( $product ) {
+
+		$should_be_synced = true;
+
+		if ( ! $this->is_product_sync_enabled() ) {
+			$should_be_synced = false;
+		}
+
+		// can't sync if we don't have a valid product object
+		if ( $should_be_synced && ! $product instanceof \WC_Product ) {
+			$should_be_synced = false;
+		}
+
+		// only published product should be synced
+		if ( $should_be_synced && 'publish' !== get_post_status( $product->get_id() ) ) {
+			$should_be_synced = false;
+		}
+
+		// make sure the given product is enabled for sync
+		if ( $should_be_synced && ! Products::product_should_be_synced( $product ) ) {
+			$should_be_synced = false;
+		}
+
+		return $should_be_synced;
 	}
 
 
@@ -4395,7 +4488,9 @@ class WC_Facebookcommerce_Integration extends WC_Integration {
 			$woo_product = new WC_Facebook_Product( $wp_id );
 		}
 
-		$products    = WC_Facebookcommerce_Utils::get_product_array( $woo_product );
+		$products = WC_Facebookcommerce_Utils::get_product_array( $woo_product );
+
+		// if the product with ID equal to $wp_id is variable, $woo_product will be the first child
 		$woo_product = new WC_Facebook_Product( current( $products ) );
 
 		// This is a generalized function used elsewhere
